move ShepherdPMF to new package - java classes
