+++ src/main/java/org/ecocean/shepherd/core/ShepherdPMF.java
@@ -1,4 +1,7 @@
-package org.ecocean;
+package org.ecocean.shepherd.core;
+
+import org.ecocean.ContextConfiguration;
+import org.ecocean.ShepherdProperties;
 
 import javax.jdo.JDOException;
 import javax.jdo.JDOHelper;
